+++ mrv2/lib/mrvGL/mrvGLViewportDraw.cpp
@@ -392,7 +392,6 @@ namespace mrv
             return;
 
         const otime::RationalTime& time = p.videoData[0].time;
-        int64_t frame = time.to_frames();
 
         const auto& annotations =
             player->getAnnotations(p.ghostPrevious, p.ghostNext);
@@ -405,9 +404,9 @@ namespace mrv
 
         for (const auto& annotation : annotations)
         {
-            int64_t annotationFrame = annotation->frame;
+            const auto& annotationTime = annotation->time;
             float alphamult = 0.F;
-            if (frame == annotationFrame || annotation->allFrames)
+            if (time == annotationTime || annotation->allFrames)
                 alphamult = 1.F;
             else
             {
@@ -415,7 +414,8 @@ namespace mrv
                 {
                     for (short i = p.ghostPrevious - 1; i > 0; --i)
                     {
-                        if (frame - i == annotationFrame)
+                        otime::RationalTime offset(i, time.rate());
+                        if (time - offset == annotationTime)
                         {
                             alphamult = 1.F - (float)i / p.ghostPrevious;
                             break;
@@ -426,7 +426,8 @@ namespace mrv
                 {
                     for (short i = 1; i < p.ghostNext; ++i)
                     {
-                        if (frame + i == annotationFrame)
+                        otime::RationalTime offset(i, time.rate());
+                        if (time + offset == annotationTime)
                         {
                             alphamult = 1.F - (float)i / p.ghostNext;
                             break;
@@ -528,7 +529,6 @@ namespace mrv
             return;
 
         const otime::RationalTime& time = p.videoData[0].time;
-        int64_t frame = time.to_frames();
 
         if (annotationsPanel)
         {
@@ -548,9 +548,9 @@ namespace mrv
 
         for (const auto& annotation : annotations)
         {
-            int64_t annotationFrame = annotation->frame;
+            const auto& annotationTime = annotation->time;
             float alphamult = 0.F;
-            if (frame == annotationFrame || annotation->allFrames)
+            if (time == annotationTime || annotation->allFrames)
                 alphamult = 1.F;
             else
             {
@@ -558,7 +558,8 @@ namespace mrv
                 {
                     for (short i = p.ghostPrevious - 1; i > 0; --i)
                     {
-                        if (frame - i == annotationFrame)
+                        otime::RationalTime offset(i, time.rate());
+                        if (time - offset == annotationTime)
                         {
                             alphamult = 1.F - (float)i / p.ghostPrevious;
                             break;
@@ -569,7 +570,8 @@ namespace mrv
                 {
                     for (short i = 1; i < p.ghostNext; ++i)
                     {
-                        if (frame + i == annotationFrame)
+                        otime::RationalTime offset(i, time.rate());
+                        if (time + offset == annotationTime)
                         {
                             alphamult = 1.F - (float)i / p.ghostNext;
                             break;