+++ mrv2/lib/mrvWidgets/mrvPlaylistButton.cpp
@@ -63,7 +63,14 @@ namespace mrv
         split_string(lines, text, "\n");
         std::string filename = lines[0] + lines[1];
         p.timeline = timeline::create(filename, context);
+        redraw();
     }
+
+    // void PlaylistButton::setTimeline(otio::Timeline* value)
+    // {
+    //     _p->timeline = value;
+    //     redraw();
+    // }
 
     int PlaylistButton::handle(int event)
     {